moved component to page directory and cleaned css
