+++ components/Profile.jsx
@@ -8,6 +8,7 @@ import { useAuthState } from "react-firebase-hooks/auth";
 
 const Profile = () => {
   const [user] = useAuthState(auth);
+  
   const fakeData = {_id:1,firstName:"Jesse",lastName:"Huang",email:"jayjay@hireup.com",freelancer:{pm:true,designer:false,engineer:true},rate:{"$numberDecimal":"95.5"},work_history:"Senior project manager for Atlantic Web Development Agency",skills:"An experienced project manager and full stack engineer",education:"Hack Reactor",location:"HRSF2",portfolio:["https://source.unsplash.com/WLUHO9A_xik/1600x900", "https://source.unsplash.com/WLUHO9A_xik/1600x900", "https://source.unsplash.com/WLUHO9A_xik/1600x900"],proposals:[{headline:" Project Sushi-Ya",overview:"Irashaimase! I'm an experienced engineer looking to create the world's most efficient salmon nigiri creation algorithm.",timeline:{start:1643514146,end:1675050146},location:"Huntington Beach, CA ",timezones:["pacific","mountain"]}],profile_pic:"https://i.ibb.co/p0fyr4S/jesse.png"}
 
   if(user) {
